Conflicts:
	sql/updates/world/12_10_29_00_kezan.sql
	sql/updates/world/12_11_08_00_creature.sql
	sql/updates/world/already_applied/12_09_13_remover linked_spell.sql
	sql/updates/world/already_applied/12_09_27_00_creature_text.sql
	sql/updates/world/already_applied/12_09_27_01_creature_text.sql
	sql/updates/world/already_applied/12_10_05_00_creature_text.sql
	sql/updates/world/already_applied/12_10_29_00_kezan.sql
	sql/updates/world/already_applied/12_11_08_00_creature.sql
	sql/updates/world/already_applied/12_11_27_00_creature.sql
	sql/updates/world/already_applied/12_12_08_kezan.sql
	src/server/authserver/Main.cpp
	src/server/game/Scripting/ScriptLoader.cpp
	src/server/scripts/EasternKingdoms/BlackwingDescent/blackwing_descent.cpp
	src/server/scripts/EasternKingdoms/BlackwingDescent/blackwing_descent.h
	src/server/scripts/EasternKingdoms/BlackwingDescent/boss_atramedes.cpp
	src/server/scripts/EasternKingdoms/BlackwingDescent/boss_chimaeron.cpp
	src/server/scripts/EasternKingdoms/BlackwingDescent/boss_magmaw.cpp
	src/server/scripts/EasternKingdoms/BlackwingDescent/boss_maloriak.cpp
	src/server/scripts/EasternKingdoms/BlackwingDescent/boss_omnotron.cpp
	src/server/scripts/EasternKingdoms/BlackwingDescent/instance_blackwing_descent.cpp
	src/server/scripts/EasternKingdoms/GrimBatol/boss_forgemaster_throngus.cpp
	src/server/scripts/EasternKingdoms/GrimBatol/grimbatol.h
	src/server/scripts/Kalimdor/ThroneOfTheFourWinds/boss_conclave_of_wind.cpp
	src/server/scripts/Kalimdor/ThroneOfTheFourWinds/instance_throne_of_the_four_winds.cpp
	src/server/scripts/Kalimdor/ThroneOfTheFourWinds/throne_of_the_four_winds.cpp
	src/server/scripts/Kalimdor/ThroneOfTheFourWinds/throne_of_the_four_winds.h

+++ src/server/game/Handlers/AuctionHouseHandler.cpp
@@ -1,11 +1,11 @@
 /*
- * Copyright (C) 2005 - 2012 MaNGOS <http://www.getmangos.com/>
+ * Copyright (C) 2005 - 2013 MaNGOS <http://www.getmangos.com/>
  *
- * Copyright (C) 2008 - 2012 Trinity <http://www.trinitycore.org/>
+ * Copyright (C) 2008 - 2013 Trinity <http://www.trinitycore.org/>
  *
- * Copyright (C) 2010 - 2012 ProjectSkyfire <http://www.projectskyfire.org/>
+ * Copyright (C) 2010 - 2013 ProjectSkyfire <http://www.projectskyfire.org/>
  *
- * Copyright (C) 2011 - 2012 ArkCORE <http://www.arkania.net/>
+ * Copyright (C) 2011 - 2013 ArkCORE <http://www.arkania.net/>
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -221,7 +221,7 @@ void WorldSession::HandleAuctionSellItem (WorldPacket & recv_data)
     AuctionHouseObject* auctionHouse = sAuctionMgr->GetAuctionsMap(pCreature->getFaction());
 
     //we have to take deposit :
-    uint32 deposit = sAuctionMgr->GetAuctionDeposit(auctionHouseEntry, etime, it, count);
+    uint64 deposit = sAuctionMgr->GetAuctionDeposit(auctionHouseEntry, etime, it, count);
     if (!pl->HasEnoughMoney(deposit))
     {
         SendAuctionCommandResult(0, AUCTION_SELL_ITEM, AUCTION_NOT_ENOUGHT_MONEY);
@@ -327,7 +327,7 @@ void WorldSession::HandleAuctionPlaceBid (WorldPacket & recv_data)
         return;
     }
 
-    if (!pl->HasEnoughMoney((uint32) price))
+    if (!pl->HasEnoughMoney((uint64) price))
     {
         //you don't have enought money!, client tests!
         //SendAuctionCommandResult(auction->auctionId, AUCTION_PLACE_BID, ???);
@@ -425,7 +425,7 @@ void WorldSession::HandleAuctionRemoveItem (WorldPacket & recv_data)
         {
             if (auction->bidder > 0)          // If we have a bidder, we have to send him the money he paid
             {
-                uint32 auctionCut = auction->GetAuctionCut();
+                uint64 auctionCut = auction->GetAuctionCut();
                 if (!pl->HasEnoughMoney(auctionCut))          //player doesn't have enough money, maybe message needed
                     return;
                 //some auctionBidderNotification would be needed, but don't know that parts..
