+++ src/server/game/Handlers/MailHandler.cpp
@@ -1,11 +1,11 @@
 /*
- * Copyright (C) 2005 - 2012 MaNGOS <http://www.getmangos.com/>
+ * Copyright (C) 2005 - 2013 MaNGOS <http://www.getmangos.com/>
  *
- * Copyright (C) 2008 - 2012 Trinity <http://www.trinitycore.org/>
+ * Copyright (C) 2008 - 2013 Trinity <http://www.trinitycore.org/>
  *
- * Copyright (C) 2010 - 2012 ProjectSkyfire <http://www.projectskyfire.org/>
+ * Copyright (C) 2010 - 2013 ProjectSkyfire <http://www.projectskyfire.org/>
  *
- * Copyright (C) 2011 - 2012 ArkCORE <http://www.arkania.net/>
+ * Copyright (C) 2011 - 2013 ArkCORE <http://www.arkania.net/>
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -113,7 +113,7 @@ void WorldSession::HandleSendMail (WorldPacket & recv_data)
 
     uint64 reqmoney = cost + money;
 
-    if (!pl->HasEnoughMoney(uint32(reqmoney)))
+    if (!pl->HasEnoughMoney(uint64(reqmoney)))
     {
         pl->SendMailResult(0, MAIL_SEND, MAIL_ERR_NOT_ENOUGH_MONEY);
         return;
@@ -421,7 +421,7 @@ void WorldSession::HandleMailTakeItem (WorldPacket & recv_data)
     }
 
     // prevent cheating with skip client money check
-    if (!pl->HasEnoughMoney(uint32(m->COD)))
+    if (!pl->HasEnoughMoney(uint64(m->COD)))
     {
         pl->SendMailResult(mailId, MAIL_ITEM_TAKEN, MAIL_ERR_NOT_ENOUGH_MONEY);
         return;
